requirements.txt
pandas
requests
pillow
python-dotenv
font-source-serif-pro
font-source-sans-pro
aggdraw
